feat: Add Deploy alias command
Folks might have a hard time finding out that they
can deploy their apps using the CLI. This change
adds an alias to `app deploy` as `deploy` under the
release section.
